requirements.txt
torch>=2.1
open_clip_torch>=2.24
numpy>=1.26
pandas>=2.0
scikit-learn>=1.3
Pillow>=10.0
tqdm>=4.66
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
